requirements.txt
httpx==0.27.0
flask==3.0.3
httpx==0.27.0
flask==3.0.3
python-dotenv==1.0.1
